TDX: fix tdx_allowed_port() and warn on disallow

+++ arch/x86/include/asm/tdx.h
@@ -88,7 +88,7 @@ bool tdx_handle_virtualization_exception(struct pt_regs *regs,
 
 bool tdx_early_handle_ve(struct pt_regs *regs);
 
-bool tdx_allowed_port(short int port);
+bool tdx_allowed_port(int port);
 
 extern phys_addr_t tdx_shared_mask(void);
